Relocated login endpoint unit tests
